[10/21] Move code from app to controllers
catalog_param_spec
